backend/requirements.txt
# ==========================================
# 🌐 CORE WEB FRAMEWORK (DJANGO & API)
# ==========================================
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# ==========================================
# 🗄️ DATABASE & VECTOR STORE (POSTGRESQL + PGVECTOR)
# ==========================================
psycopg2-binary>=2.9.9      # Driver để Django kết nối với PostgreSQL
pgvector>=0.2.5             # Thư viện hỗ trợ kiểu dữ liệu Vector trong Postgres
langchain-postgres>=0.0.9   # Để Langchain lưu vector thẳng vào Postgres
faiss-cpu>=1.8.0            # Vẫn giữ FAISS dự phòng hoặc dùng cho test local

# ==========================================
# 🧠 AI & LANGCHAIN CORE
# ==========================================
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.0.1
langchain-text-splitters>=0.0.1

# ==========================================
# 📄 DOCUMENT LOADERS (XỬ LÝ TÀI LIỆU)
# ==========================================
pymupdf>=1.23.0             # Máy cày đọc PDF siêu việt (thay cho pypdf)
docx2txt>=0.8               # Đọc file Word (DOCX)
unstructured>=0.12.0        # Xử lý hình ảnh và các định dạng phức tạp

# ==========================================
# 🔒 SECURITY & AUTH (TÙY CHỌN)
# ==========================================
# Mở comment dòng dưới nếu team code xong phần Đăng nhập/Đăng ký
djangorestframework-simplejwt>=5.3.1

# ==========================================
# 👁️ OCR & IMAGE PROCESSING (MỚI THÊM CHO RAG)
# ==========================================
paddleocr>=2.7.0.3          # Lõi OCR đọc chữ từ ảnh/PDF scan
paddlepaddle>=2.6.0         # Engine bắt buộc để chạy được paddleocr
opencv-python>=4.8.0        # Xử lý, lọc nhiễu ảnh (cv2)
Pillow>=10.2.0              # Xử lý định dạng ảnh (PIL)
numpy>=1.26.0               # Tính toán ma trận ảnh cho OpenCV

# ==========================================
# 🛠️ UTILS & HTTP REQUESTS (GIAO TIẾP FE - BE)
# ==========================================
requests>=2.31.0            # Cực kỳ quan trọng để Streamlit gọi API xuống Django
python-dotenv>=1.0.0        # Đọc các biến môi trường bảo mật từ file .env

# ==========================================
# 🧩 LANGCHAIN CLASSIC (HỖ TRỢ QUERY CHAINS)
# ==========================================
langchain-classic>=1.0.0    # Chứa các module create_retrieval_chain cũ nhưng rất ổn định

# ==========================================
# 📝 MỞ RỘNG XỬ LÝ TÀI LIỆU (CÂU HỎI 1)
# ==========================================
python-docx>=1.1.0          # Lõi đọc và xử lý file Word (.docx) chuyên sâu